Move tree.py out of util directory
